drm/clearkey: don't require DataSize of InitData to be 0
The spec says:
"cenc" Initialization Data Format - 4.1

"Data and DataSize"
Reserved for future use. DataSize shall be set to 0 when constructing
this box. When processing, if dataSize is non-zero the Data field
shall be ignored.

https://www.w3.org/TR/eme-initdata-cenc/#common-system-definition

Test: In real life we have met a media file with clearkey pssh
that has DataSize != 0, and it was causing the plugin to refuse its
initialization. Unfortunatelly I can't share it.

Change-Id: I300804e6040b1363646514e4b09258f274da8da6